add initializeAndLoadQiitaDir methods and rename to qiita-settings
